+++ packages/analytics/dashboard-renderer/src/components/DashboardRenderer.cy.ts
@@ -25,6 +25,7 @@ import {
   summaryDashboardConfig,
   simpleConfigNoFilters,
   fourByFourDashboardConfigJustCharts, oneTileDashboardConfig,
+  simpleConfigGlobalFilters,
 } from '../../sandbox/mock-data'
 import { createPinia, setActivePinia } from 'pinia'
 import { EntityLink } from '@kong-ui-public/entities-shared'
@@ -134,10 +135,6 @@ describe('<DashboardRenderer />', () => {
         },
       },
       modelValue: {
-        gridSize: {
-          cols: 2,
-          rows: 4,
-        },
         tiles: [
           {
             definition: {
@@ -379,7 +376,6 @@ describe('<DashboardRenderer />', () => {
         timeSpec: customTimeframe.v4Query(),
       },
       modelValue: {
-        gridSize: { cols: 3, rows: 2 },
         tiles: [
           {
             definition: {
@@ -433,7 +429,6 @@ describe('<DashboardRenderer />', () => {
         timeSpec: customTimeframe.v4Query(),
       },
       modelValue: {
-        gridSize: { cols: 3, rows: 2 },
         tiles: [
           {
             definition: {
@@ -650,6 +645,54 @@ describe('<DashboardRenderer />', () => {
     })
   })
 
+  it('merges global filters', () => {
+    const props = {
+      context: {
+        filters: [
+          // Valid filter
+          {
+            field: 'api_product',
+            operator: 'in',
+            value: ['some product'],
+          },
+        ],
+      },
+      modelValue: simpleConfigGlobalFilters,
+    }
+
+    cy.mount(DashboardRenderer, {
+      props,
+      global: {
+        provide: {
+          [INJECT_QUERY_PROVIDER]: mockQueryProvider(),
+        },
+      },
+    }).then(() => {
+      // Extra calls may mean we mistakenly issued queries before knowing the timeSpec.
+      cy.get('@fetcher').should('have.callCount', 3)
+      cy.get('@fetcher').should('always.have.been.calledWithMatch', Cypress.sinon.match({
+        datasource: 'advanced',
+        query: {
+          filters: [
+            {
+              field: 'api_product',
+              operator: 'in',
+              value: ['some product'],
+            },
+            {
+              field: 'control_plane',
+              operator: 'in',
+              value: ['default_uuid'],
+            },
+          ],
+        },
+      }))
+
+      // Check that it replaces the description token.
+      cy.get('.header-description').should('have.text', 'Last 7-Day Summary')
+    })
+  })
+
   it('renders tile timeframe overrides in a badge', () => {
     const props = {
       context: {
@@ -660,10 +703,6 @@ describe('<DashboardRenderer />', () => {
         },
       },
       modelValue: {
-        gridSize: {
-          cols: 6,
-          rows: 4,
-        },
         tileHeight: 167,
         tiles: [
           {
@@ -871,7 +910,6 @@ describe('<DashboardRenderer />', () => {
   })
 
   it('tiles maintain row-column order after reordering', () => {
-
     const configRef = ref<DashboardConfig>(fourByFourDashboardConfigJustCharts)
     const props = {
       context: {
@@ -909,121 +947,6 @@ describe('<DashboardRenderer />', () => {
     cy.wrap(configRef).should((ref: Ref<DashboardConfig>) => {
       const currentOrder = ref.value.tiles.map((tile: TileConfig) => tile.id)
       expect(currentOrder).to.deep.equal(updatedTileIDOrder)
-    })
-  })
-
-  it('Update gridsize when tile added', () => {
-    const configRef = ref<DashboardConfig>(fourByFourDashboardConfigJustCharts)
-    const props = {
-      context: {
-        filters: [],
-        timeSpec: {
-          type: 'relative',
-          time_range: '15m',
-        },
-        editable: true,
-      },
-      modelValue: configRef.value,
-    }
-
-    cy.mount(DashboardRenderer, {
-      props,
-      global: {
-        provide: {
-          [INJECT_QUERY_PROVIDER]: mockQueryProvider(),
-        },
-      },
-    })
-
-    expect(configRef.value.gridSize).to.deep.equal({ cols: 8, rows: 6 })
-
-    cy.wrap(configRef).should((ref: Ref<DashboardConfig>) => {
-      ref.value.tiles.push({
-        id: crypto.randomUUID(),
-        definition: {
-          chart: {
-            type: 'timeseries_line',
-            chartTitle: 'New Tile',
-          },
-          query: {
-            metrics: ['request_count'],
-            dimensions: ['time'],
-            filters: [],
-          },
-        },
-        layout: {
-          position: {
-            col: 0,
-            row: 0,
-          },
-          size: {
-            cols: 4,
-            rows: 4,
-          },
-        },
-      })
-    })
-
-    cy.wrap(configRef).should((ref: Ref<DashboardConfig>) => {
-      expect(ref.value.gridSize).to.deep.equal({ cols: 8, rows: 8 })
-    })
-  })
-
-  it('Preserve column counts', () => {
-    const configRef = ref<DashboardConfig>(oneTileDashboardConfig)
-
-    const props = {
-      context: {
-        filters: [],
-        timeSpec: {
-          type: 'relative',
-          time_range: '15m',
-        },
-        editable: true,
-      },
-      modelValue: configRef.value,
-    }
-
-    cy.mount(DashboardRenderer, {
-      props,
-      global: {
-        provide: {
-          [INJECT_QUERY_PROVIDER]: mockQueryProvider(),
-        },
-      },
-    })
-
-    expect(configRef.value.gridSize).to.deep.equal({ cols: 6, rows: 2 })
-
-    cy.wrap(configRef).should((ref: Ref<DashboardConfig>) => {
-      ref.value.tiles.push({
-        id: crypto.randomUUID(),
-        definition: {
-          chart: {
-            type: 'timeseries_line',
-            chartTitle: 'New Tile',
-          },
-          query: {
-            metrics: ['request_count'],
-            dimensions: ['time'],
-            filters: [],
-          },
-        },
-        layout: {
-          position: {
-            col: 0,
-            row: 2,
-          },
-          size: {
-            cols: 4,
-            rows: 4,
-          },
-        },
-      })
-    })
-
-    cy.wrap(configRef).should((ref: Ref<DashboardConfig>) => {
-      expect(ref.value.gridSize).to.deep.equal({ cols: 6, rows: 6 })
     })
   })
 })