+++ packages/entities/entities-vaults/src/components/VaultForm.vue
@@ -499,6 +499,17 @@
               show-password-mask-toggle
               type="password"
             />
+            <KCheckbox
+              v-if="config.base64FieldAvailable"
+              v-model="configFields[VaultProviders.CONJUR].base64_decode!"
+              data-testid="vault-form-config-env-base64_decode"
+              :label="t('form.config.commonFields.base64_decode.label')"
+              :label-attributes="{
+                info: t('form.config.commonFields.base64_decode.tooltip'),
+                tooltipAttributes: { maxWidth: '400' },
+              }"
+              :readonly="form.isReadonly"
+            />
           </div>
 
           <div v-if="config.ttl">
@@ -811,7 +822,8 @@ const configFields = reactive<ConfigFields>({
   } as AzureVaultConfig,
   [VaultProviders.CONJUR]: {
     endpoint_url: '',
-    auth_method: 'default',
+    auth_method: 'api_key',
+    ...base64FieldConfig,
   },
 })
 
@@ -862,7 +874,8 @@ const originalConfigFields = reactive<ConfigFields>({
   } as AzureVaultConfig,
   [VaultProviders.CONJUR]: {
     endpoint_url: '',
-    auth_method: 'default',
+    auth_method: 'api_key',
+    ...base64FieldConfig,
   },
 })